% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsAccessors.R
\name{cellCountsExpr}
\alias{cellCountsExpr}
\title{Return the total number of genes expressed for each cell}
\usage{
cellCountsExpr(TenXSubset)
}
\arguments{
\item{TenXSubset}{A TenXSubset object}
}
\value{
A numeric vector giving the total number of genes expressed for each
cell.
}
\description{
A quick method to access the \code{expressed} column of the
\code{cellCounts} data frame.
}
