% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsAccessors.R
\name{cellCountsTrans}
\alias{cellCountsTrans}
\title{Return the total number of transcripts for each cell}
\usage{
cellCountsTrans(TenXSubset)
}
\arguments{
\item{TenXSubset}{A TenXSubset object}
}
\value{
A numeric vector giving the transcript counts for each cell.
}
\description{
A quick method to access the \code{transcripts} column of the
\code{cellCounts} data frame.
}
