% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsAccessors.R
\name{cellGroup}
\alias{cellGroup}
\title{Return the cellGroup vector}
\usage{
cellGroup(TenXSubset)
}
\arguments{
\item{TenXSubset}{A TenXSubset object}
}
\value{
A numeric vector giving the ID of the group each cell belongs to.
}
\description{
A quick method to access the cellGroup that each cell belongs to.
}
\details{
Since we bin the cells into small groups when we want to look at the data as
a whole (and one cannot do that due to memory restrictions unless summarised
data is used), this is useful for determining which group should be removed
if we find that a particular cell does not fit in with the rest.
}
