% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsAccessors.R
\name{cellIndex}
\alias{cellIndex}
\title{Cells in data}
\usage{
cellIndex(TenXSubset)
}
\arguments{
\item{TenXSubset}{A TenXSubset object}
}
\value{
A numeric vector of indices.
}
\description{
Return cell indices of the cells included in this TenXSubset object
relative to the original, full 10x Genomics 1.3 Million cell data set.
}
