% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsAccessors.R
\name{cellLib}
\alias{cellLib}
\title{Return the libraryID vector}
\usage{
cellLib(TenXSubset)
}
\arguments{
\item{TenXSubset}{A TenXSubset object}
}
\value{
A numeric vector giving the ID of the library each cell belongs to.
}
\description{
A quick method to access the library that each cell belongs to.
}
