% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsAccessors.R
\name{cellMouse}
\alias{cellMouse}
\title{Return the mouseID vector}
\usage{
cellMouse(TenXSubset)
}
\arguments{
\item{TenXSubset}{A TenXSubset object}
}
\value{
A character vector of A's and B's characterising each cell by mouse.
}
\description{
A quick method to access the mouse (A or B) associated with each cell in the
data.
}
