% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsAccessors.R
\name{cellN}
\alias{cellN}
\title{Total number of cells in the data set}
\usage{
cellN(TenXSubset)
}
\arguments{
\item{TenXSubset}{A TenXSubset object}
}
\description{
Total number of cells in the data set
}
