% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsIndexing.R
\name{determineGeneFromGO}
\alias{determineGeneFromGO}
\title{Find genes associated with a GO term}
\usage{
determineGeneFromGO(goTerm)
}
\arguments{
\item{goTerm}{The GO id of interest.}
}
\value{
A \code{data.frame} of:
\itemize{
    \item HGNC gene names
    \item ENSEMBL gene ids
    \item indices of these genes in our dataset (optional)
}
}
\description{
Given a GO term, find the genes annotated with this GO term.
}
