% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsIndexing.R
\name{determineGeneFromQuery}
\alias{determineGeneFromQuery}
\title{Find genes containing some query in their GO description}
\usage{
determineGeneFromQuery(query)
}
\arguments{
\item{query}{A character string of interest included in the gene decription.}
}
\value{
A list of genes with this particular phrase in one of their
        descriptions.
}
\description{
Given some query/phrase, find the genes which include this in their GO
annotation or description.
}
