% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsIndexing.R
\name{determineGeneID}
\alias{determineGeneID}
\title{Find gene name in data}
\usage{
determineGeneID(geneName, TenXSubset)
}
\arguments{
\item{TenXSubset}{A \code{TenXSubset} object.}

\item{queryGroupIndexVector}{A numeric vector giving the group indices we
would like to retrieve cell/sample indices from.}
}
\value{
A numeric index of the gene in the data set if the gene is found. The
        function throws an error otherwise.
}
\description{
Given a HGNC gene name or ENSEMBL id, find the index of this gene in the data
set. This function is not case-sensitive.
}
