% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsIndexing.R
\name{determineGOFromGene}
\alias{determineGOFromGene}
\title{Find GO terms associated with a gene}
\usage{
determineGOFromGene(geneName)
}
\arguments{
\item{geneName}{A character string giving either the HGNC representation of
the gene or the ENSEMBL id.}
}
\value{
A \code{data.frame} detailing the GO annotations associated with the
        chosen gene/genes.
}
\description{
Given a gene HGNC name or EMSEMBL id, find the GO description associated with
this gene. This is very useful when analysing outliers in the data.
}
