% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsIndexing.R
\name{determineGroupAnnotations}
\alias{determineGroupAnnotations}
\title{Determine group annotations}
\usage{
determineGroupAnnotations(groupSize, cellN, cellMouse)
}
\arguments{
\item{groupSize}{The size of the bin the cells should be grouped by.}

\item{cellN}{The total number of cells to consider.}

\item{cellMouse}{The annotation vector partitioning the cells into mouse A
and B.}
}
\value{
A vector of group indices as long as \code{cellN}.
}
\description{
Although this is generally calculated when a new \code{TenXSubset} object is
created, sometimes we with to annotate the groups from scratch. This function
allows one to annotate the samples into groups of a specified size (using the
mouse labeling such that there is no group with two mice).
}
