% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsIndexing.R
\name{determineLibraryAnnotations}
\alias{determineLibraryAnnotations}
\title{Determine library annotations}
\usage{
determineLibraryAnnotations(groupSize, cellLib, cellN, cellMouse)
}
\arguments{
\item{groupSize}{The size of the bin the cells should be grouped by.}

\item{cellLib}{The library each of the cells belongs to.}

\item{cellN}{The total number of cells to consider.}

\item{cellMouse}{The annotation vector partitioning the cells into mouse A
and B.}
}
\value{
A vector of group indices as long as \code{cellN}.
}
\description{
Function required by \code{determineLibraryFromGroup} from matches each group
with a library (based on majority counts).
}
