% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsIndexing.R
\name{determineLibraryFromGroup}
\alias{determineLibraryFromGroup}
\title{Get the library that this group belongs to}
\usage{
determineLibraryFromGroup(queryGroupIndexVector, TenXSubset)
}
\arguments{
\item{queryGroupIndexVector}{A numeric vector giving the group indices we
would like to retrieve cell/sample indices from.}

\item{TenXSubset}{A \code{TenXSubset} object.}
}
\value{
An index vector identifying the cells in this group(s).
}
\description{
Given a group index, determine the library the group comes from. In some
cases groups can be spread over multiple libraries. In this case the group is
assigned to the "majority" library.
}
