% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsIndexing.R
\name{determineMouseFromGroup}
\alias{determineMouseFromGroup}
\title{Get the mouse that this group belongs to}
\usage{
determineMouseFromGroup(queryGroupIndexVector, TenXSubset)
}
\arguments{
\item{queryGroupIndexVector}{A numeric vector giving the group indices we
would like to retrieve cell/sample indices from.}

\item{TenXSubset}{A \code{TenXSubset} object.}
}
\value{
A character vector identifying the mouse associated with each of
        these groups.
}
\description{
Given a group index, determine the mouse that this group of cells came from.
}
