% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsAccessors.R
\name{geneCounts}
\alias{geneCounts}
\title{Return geneCounts data.frame}
\usage{
geneCounts(TenXSubset)
}
\arguments{
\item{TenXSubset}{A TenXSubset object}
}
\value{
\code{geneCounts} data frame.
}
\description{
A quick method to access the geneCounts \code{data.frame} within a TenXSubset
object.
}
\details{
The \code{data.frame} contains the transcript counts associated with each
gene for each group in the data set. If this method is used (rather than
simply viewing the contents of the \code{geneCounts} slot), the gene names
will automatically be included as row names.
}
