% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsAccessors.R
\name{geneENSEMBL}
\alias{geneENSEMBL}
\title{Return the ENSEMBL names of the genes}
\usage{
geneENSEMBL(TenXSubset)
}
\arguments{
\item{TenXSubset}{A TenXSubset.}
}
\value{
A character vector of ENSEMBL gene ids.
}
\description{
A quick method to access which ENSEMBL ids are included in this TenXSubset.
}
