% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsAccessors.R
\name{geneExpressed}
\alias{geneExpressed}
\title{Return geneExpressed data.frame}
\usage{
geneExpressed(TenXSubset)
}
\arguments{
\item{TenXSubset}{A TenXSubset object}
}
\value{
\code{geneExpressed} data frame.
}
\description{
A quick method to access the geneExpressed \code{data.frame} within a
TenXSubset object.
}
\details{
The \code{data.frame} contains the proportion of times each gene is expressed
in each group of cells. If this method is used (rather than simply viewing
the contents of the \code{geneExpressed} slot), the gene names will
automatically be included as row names.
}
