% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsAccessors.R
\name{geneIndex}
\alias{geneIndex}
\title{Genes in data}
\usage{
geneIndex(TenXSubset)
}
\arguments{
\item{TenXSubset}{A TenXSubset object}
}
\value{
A numeric vector of indices.
}
\description{
Return gene indices of the cells included in this TenXSubset object
relative to the original, full 10x Genomics 1.3 Million cell data set.
}
