% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsAccessors.R
\name{geneN}
\alias{geneN}
\title{Total number of genes in the data set}
\usage{
geneN(TenXSubset)
}
\arguments{
\item{TenXSubset}{A TenXSubset object}
}
\description{
Total number of genes in the data set
}
