% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsAccessors.R
\name{geneNames}
\alias{geneNames}
\title{Return the HGNC names of the genes}
\usage{
geneNames(TenXSubset)
}
\arguments{
\item{TenXSubset}{A TenXSubset.}
}
\value{
A character vector of gene names.
}
\description{
A quick method to access which genes are included in this TenXSubset.
}
