% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{generateSubset}
\alias{generateSubset}
\title{Generate a TenXSubset object}
\usage{
generateSubset(cellIndex, geneIndex, groupSize, path)
}
\arguments{
\item{cellIndex}{a numeric vector giving the cells (with reference to the
original data set) that should be included in this subset}

\item{geneIndex}{a numeric vector giving the genes that should be included
in this subset}

\item{groupSize}{the number of cells which should be included in each summary
bin}

\item{path}{a character string which specifies the path to the .h5 TenX
object which was previously downloaded}
}
\value{
A \code{TenXSubset} object. See \code{help(TenXSubset-class)} for more
        details.
}
\description{
Generate a TenXSubset object
}
