% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{generateSummaryExpr}
\alias{generateSummaryExpr}
\title{Generate a summary counts object}
\usage{
generateSummaryExpr(TenXSubset)
}
\arguments{
\item{object}{A \code{TenXSubset} object with groupSize = 1.}
}
\value{
A data.frame giving the cell, library, mouse, and gene counts
        associated with each gene which can then be used in the
        \code{plotTenXMouseExpr} and ... functions.
}
\description{
Generate a summary counts object
}
