% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsAccessors.R
\name{groupN}
\alias{groupN}
\title{Total number of groups in the data set}
\usage{
groupN(TenXSubset)
}
\arguments{
\item{TenXSubset}{A TenXSubset object}
}
\description{
Total number of groups in the data set
}
