% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsAccessors.R
\name{groupSize}
\alias{groupSize}
\title{Size of summary groups}
\usage{
groupSize(TenXSubset)
}
\arguments{
\item{TenXSubset}{A TenXSubset object}
}
\value{
A numeric vector of indices.
}
\description{
Return the size of bins used to group the cells together. E.g. if this is
100 then each group annotated in \code{geneCounts} will have summary counts
over 100 cells.
}
\details{
Note that when the cells are being partitioned, we ensure that the mice do
not overlap, thus there will usually be two groups with less than
\code{groupSize} number of cells.
}
