% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsAccessors.R
\name{libN}
\alias{libN}
\title{Total number of libraries in the data set}
\usage{
libN(TenXSubset)
}
\arguments{
\item{TenXSubset}{A TenXSubset object}
}
\description{
Total number of libraries in the data set
}
