% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsPlotting.R
\name{plotTenX}
\alias{plotTenX}
\title{TenXSubset object data summaries}
\usage{
plotTenX(object, x, y, type, colorByCells = NULL, colorByGenes = NULL, ...)
}
\arguments{
\item{object}{A \code{TenXSubset} object.}

\item{x}{A \code{character} string indicating the values to appear on the x
axis.}

\item{y}{An optional \code{character} vector indicating the values to appear
on the y axis.}

\item{type}{An optional \code{character} string representing the type of
figure to plot. This is one of \code{histogram}, \code{density},
\code{boxplot}, \code{violin}, or \code{point}. If this is left blank, the
algorithm will output a list of suitable choices for the user  to choose
based on the \code{x} and \code{y} parameters.}

\item{colorByCells}{An optional parameter specifying either:
\itemize{
    \item the indices of the cells to be coloured differently (end up with
    two distinct groups)
    \item \code{"library"} or \code{"mouse"} as a shortcut to colour based on
    existing groups
}}

\item{colorByGenes}{An optional parameter specifying the \code{numeric}
indices of the genes to be coloured differently.}

\item{...}{Any additional parameters compatible with \pkg{ggplot2}.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
A function that plots the data summaries with the quantities of
interest defined by the user. See Details for the allowed combinations.
}
\details{
The basic idea of this function is to provide an each interface which allows
one to visualise the data in different meaningful ways in order to locate
inconsistencies or possible problems which may arise.

There are multiple combinations of parameters one can use, but obviously not
all will work well together. As a result, the function provides the user with
prompts:
\itemize{
    \item if necessary parameters are left out, the function suggests
    suitable values for these parameters
    \item if the parameters chosen are not compatible, the function suggests
    alternative parameter values based on the ones that have been chosen
}
}
