% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsPlotting.R
\name{plotTenXGene}
\alias{plotTenXGene}
\title{Plot gene summary for TenXSubset object}
\usage{
plotTenXGene(object, geneName, ...)
}
\arguments{
\item{object}{A \code{TenXSubset} object.}

\item{geneName}{A HGNC or ENSEMBL gene identifier. Note that only those nemes
included in \code{geneNames(TenXSubset)} or
\code{geneENSEMBL(TenXSubset)} can be used.}
}
\value{
Visualise how common the gene is compared to the others in the data
set. The returns a table of 4 values giving the:
\itemize{
    \item name of the gene
    \item number of transcripts of this gene in this dataset
    \item number of cells that express this gene in this dataset
    \item number of transcripts per cell expressing this gene
}
}
\description{
Plot gene summary for TenXSubset object
}
