% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsPlotting.R
\name{plotTenXGeneCor}
\alias{plotTenXGeneCor}
\title{Compare gene expression across multiple genes}
\usage{
plotTenXGeneCor(object, geneName)
}
\arguments{
\item{object}{A \code{TenXSubset} object.}

\item{geneName}{A vector of HGNC or ENSEMBL gene names.}
}
\description{
Compare gene expression across multiple genes
}
