% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsPlotting.R
\name{plotTenXGeneGO}
\alias{plotTenXGeneGO}
\title{Plot GO summary for specified gene set}
\usage{
plotTenXGeneGO(object, geneName)
}
\arguments{
\item{object}{A \code{TenXSubset} object.}

\item{geneName}{A vector of HGNC or ENSEMBL gene names.}
}
\description{
Given a set of gene indices, determine the top 10 functions associated with
these genes and plot them as a bar chart.
}
