% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsPlotting.R
\name{plotTenXGenePair}
\alias{plotTenXGenePair}
\title{Compare gene expression profiles for two genes}
\usage{
plotTenXGenePair(object, geneName1, geneName2)
}
\arguments{
\item{object}{A \code{TenXSubset} object.}

\item{geneName1}{A HGNC or ENSEMBL gene identifier.}

\item{geneName2}{A HGNC or ENSEMBL gene identifier.}
}
\description{
Compare gene expression profiles for two genes
}
