% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsPlotting.R
\name{plotTenXGO}
\alias{plotTenXGO}
\title{Plot GO summary for TenXSubset object}
\usage{
plotTenXGO(object, goId, plotAll, ...)
}
\arguments{
\item{object}{A \code{TenXSubset} object.}

\item{goId}{A valid GO id.}

\item{plotAll}{Whether the rankings should be plotted or not (if \code{TRUE}
then this is the only output).}
}
\description{
Plot GO summary for TenXSubset object
}
