% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsPlotting.R
\name{plotTenXGroupExpr}
\alias{plotTenXGroupExpr}
\title{Plot how gene expression changes based on library}
\usage{
plotTenXGroupExpr(object, geneName)
}
\arguments{
\item{object}{A \code{TenXSubset} object.}

\item{geneName}{A vector of HGNC or ENSEMBL gene names.}
}
\description{
Plot how gene expression changes based on library
}
