% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericsPlotting.R
\name{plotTenXMouseExpr}
\alias{plotTenXMouseExpr}
\title{Plot how gene expression changes based on mouse}
\usage{
plotTenXMouseExpr(summaryExpr, geneName, plot)
}
\arguments{
\item{summaryExpr}{A data.frame giving the cell index, library, mouse, and
transcript counts associated with each of a group of cells.}

\item{geneName}{A HGNC or ENSEMBL gene name that corresponds to one of the
column names of \code{summaryExpr}.}

\item{plot}{One of "all", "point", "hist", or "boxplot".}
}
\description{
Given a data.frame generated by \code{generateSummaryExpr}, illustrate how
gene expression changes across mice for a specific gene.
}
