% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\name{TenXse_colData}
\alias{TenXse_colData}
\title{colData of the original data set}
\description{
A \code{data.frame} containing:
\itemize{
    \item The \code{Barcode} associated with each cell (a combination of a
    sequence and the library number)
    \item The \code{Sequence} identifying this particular cell.
    \item The \code{Library} that this cell belongs to.
    \item The \code{Mouse} that this cell belongs to (either A or B)
}
}
