% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\name{TenXse_rowData}
\alias{TenXse_rowData}
\title{rowData of the original data set}
\description{
A data.frame containing the gene names (HGNC) and ENSEMBL id's of each gene
included in the original data set.
}
