% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TenXSubset-class}
\alias{TenXSubset-class}
\title{Class used to represent the grouped TenXData. The object provides summarised
information about the chosen subset of cells which can then be analysed using
a variety of plotting functions.}
\description{
Class used to represent the grouped TenXData. The object provides summarised
information about the chosen subset of cells which can then be analysed using
a variety of plotting functions.
}
\section{Slots}{

\describe{
\item{\code{geneCounts}}{A data.frame with genes as rows and sample groups as
columns. Each element represents the total number of transcripts associated
with the gene in the chosen cell group.}

\item{\code{geneExpressed}}{Similar to geneCounts, here we determine the number of cells
in which this particular gene is expressed (out of the total number in the
group).}

\item{\code{geneNames}}{The names of the genes associated with each row.}

\item{\code{geneENSEMBL}}{The ENSEMBL id associated with each row.}

\item{\code{cellCounts}}{A data.frame which documents the number of transcripts per
cell and number of unexpressed genes per cell. These can be accessed through
"cellTranscripts" and "cellExpressed"}

\item{\code{cellMouse}}{A vector giving the mouse associated with each cell.}

\item{\code{cellLib}}{The library associated with each cell in the data.}

\item{\code{cellGroup}}{The group associated with each cell in the data,
characterised by \code{groupSize}.}

\item{\code{cellN}}{The number of cells included in this data set.}

\item{\code{groupN}}{The number of distinct groups of those samples.}

\item{\code{libN}}{The number of distinct libraries to which the cells belong.}

\item{\code{geneN}}{The number of genes for which we have associated counts.}

\item{\code{cellIndex}}{The cells used in this subset relative to the original
1.3 million cell data.}

\item{\code{geneIndex}}{The genes included in this subset.}

\item{\code{groupSize}}{The size of the bins used to summarise the data.}
}}

